+++ artemis/src/test/java/com/artemis/EntitySubscriptionTest.java
@@ -1,9 +1,8 @@
 package com.artemis;
 
+import com.artemis.annotations.Wire;
 import com.artemis.component.ComponentX;
 import com.artemis.component.ComponentY;
-import com.artemis.systems.EntityProcessingSystem;
-import com.artemis.utils.ImmutableBag;
 import com.artemis.utils.IntBag;
 import org.junit.Test;
 
@@ -60,8 +59,20 @@ public class EntitySubscriptionTest {
 		assertEquals(3, sm.removed);
 	}
 
+	@Test
+	public void subscription_remove_id_matches_entity() {
+		WorldConfiguration config = new WorldConfiguration();
+		config.setManager(new TestManager());
+		World world = new World(config);
 
-	private static class SubscribingManager extends Manager
+		Entity entity = world.createEntity();
+		world.process();
+		entity.deleteFromWorld();
+		world.process();
+	}
+
+	@Wire
+	static class SubscribingManager extends Manager
 			implements  EntitySubscription.SubscriptionListener {
 
 		int inserted, removed;
@@ -93,4 +104,37 @@ public class EntitySubscriptionTest {
 			removed += entities.size();
 		}
 	}
+
+		@Wire
+	static class TestManager extends Manager {
+			AspectSubscriptionManager subscriptionManager;
+
+			@Override
+			protected void initialize() {
+				EntitySubscription subscription = subscriptionManager.get(Aspect.all());
+
+				subscription.addSubscriptionListener(new EntitySubscription.SubscriptionListener() {
+					@Override
+					public void inserted(IntBag entities) {
+						int[] data = entities.getData();
+						for (int i = 0; i < entities.size(); i++) {
+							int entityId = data[i];
+							assertEquals(0, entityId);
+
+							assertNotNull(world.getEntity(entityId));
+						}
+					}
+
+					@Override
+					public void removed(IntBag entities) {
+						int[] data = entities.getData();
+						for (int i = 0; i < entities.size(); i++) {
+							int entityId = data[i];
+							assertEquals(0, entityId);
+							assertNotNull(world.getEntity(entityId));
+						}
+					}
+				});
+			}
+		}
 }