+++ artemis/src/test/java/com/artemis/WorldTest.java
@@ -2,6 +2,7 @@ package com.artemis;
 
 import static org.junit.Assert.assertEquals;
 
+import com.artemis.utils.IntBag;
 import org.junit.Assert;
 import org.junit.Test;